+++ index.d.ts
@@ -1,0 +1,3 @@
+// index.d.ts
+declare module "*.jpg";
+declare module "*.jpeg";
