+++ src/pages/login-page/LoginReg.styles.ts
