+++ src/pages/main-page/MainPage.tsx
@@ -8,23 +8,24 @@ import { Centerblock } from '../../components/Centerblock/Centerblock'
 import { logInState } from '../../store/actions/creators/authCreator'
 import * as S from '../../App.styles'
 
+
 export const Main = () => {
-  const trackInPlayer = useSelector(currentTrackSelector)
-  const userInfo = JSON.parse(localStorage.getItem('userSkyproMusic'))
-  const dispatch = useDispatch()
-  useEffect(() => {
-    if (userInfo) {
-      dispatch(logInState(userInfo))
-    }
-  }, [])
-  return (
-    <>
-      <S.Main>
-        <NavMenu />
-        <Centerblock />
-        <Sidebar />
-      </S.Main>
-      {trackInPlayer && <BarPlayer />}
-    </>
-  )
+	const trackInPlayer = useSelector(currentTrackSelector)
+	const userInfo = JSON.parse(localStorage.getItem('userSkyproMusic') || '{}')
+	const dispatch = useDispatch()
+	useEffect(() => {
+		if (userInfo) {
+			dispatch(logInState(userInfo))
+		}
+	}, [])
+	return (
+		<>
+			<S.Main>
+				<NavMenu />
+				<Centerblock />
+				<Sidebar />
+			</S.Main>
+			{trackInPlayer && <BarPlayer />}
+		</>
+	)
 }
