--- src/store/selectors/audioplayerSelectors.js
@@ -1,17 +1,0 @@
-const audioplayerSelector = (store) => store.audioplayer
-
-// список треков с API
-export const playListSelector = (store) => audioplayerSelector(store)?.playlist
-
-// перемешанный список треков с API
-export const playListShuffleSelector = (store) =>
-  audioplayerSelector(store).shuffledPlaylist
-
-// текущий трек в плеере
-export const currentTrackSelector = (store) => audioplayerSelector(store)?.track
-
-// статус воспроизведения
-export const isPlauingSelector = (store) => audioplayerSelector(store)?.plauing
-
-// статус воспроизведения в перемешку
-export const isShuffledSelector = (store) => audioplayerSelector(store).shuffled
