+++ tsconfig.json
@@ -1,0 +1,13 @@
+{
+	"compilerOptions": {
+	  "strict": true, // Включает все жесткие проверки типов
+	  "outDir": "./dist/",
+	  "noImplicitAny": true,
+	  "module": "es6",
+	  "target": "es5",
+	  "allowJs": true,
+	  "jsx": "react-jsx",
+	  "moduleResolution": "node",
+	  "esModuleInterop": true,
+	}
+ }
